BE: deletet wishlist and register endpoint of Event and combined functionality with wishlist and register endpoint of User
