+++ backend/src/controller/index.js
@@ -9,12 +9,10 @@ import { getAllEventsCtrl } from "./eventController.js";
 import { getEventDetailsCtrl } from "./eventController.js";
 import { getUserProfileInfoCtrl } from "./userController.js";
 import { patchEventToWishlistCtrl } from "./userController.js";
-import { patchFillRegisteredGuestCheckFullyBookedCtrl } from "./eventController.js";
 import { postFileUploadCtrl } from "./fileUploadController.js";
 import { patchEditUserProfileCtrl } from "./userController.js";
 import { patchEventToRegisteredEventsListCtrl } from "./userController.js";
 import { patchEventIsCancelledCtrl } from "./eventController.js";
-import { patchFillWishlistCounterCtrl } from "./eventController.js";
 import { patchEditEventCtrl } from "./eventController.js";
 import { postNewReviewCtrl } from "./reviewController.js";
 import { getHostDetailsCtrl } from "./userController.js";
@@ -35,8 +33,6 @@ export const EventController = {
     postNewEventCtrl,
     getAllEventsCtrl,
     getEventDetailsCtrl,
-    patchFillRegisteredGuestCheckFullyBookedCtrl,
-    patchFillWishlistCounterCtrl,
     patchEditEventCtrl,
     patchEventIsCancelledCtrl,
 };
