+++ backend/src/routes/eventRouter.js
@@ -8,20 +8,9 @@ export const eventRouter = express
     .get("/", doJwtAuth, EventController.getAllEventsCtrl)
     .get("/:eventId", doJwtAuth, EventController.getEventDetailsCtrl)
     .post("/", doJwtAuth, EventController.postNewEventCtrl)
-    .patch(
-        "/register/:eventId",
-        doJwtAuth,
-        EventController.patchFillRegisteredGuestCheckFullyBookedCtrl
-    )
-    .patch(
-        "/wishlist/:eventId",
-        doJwtAuth,
-        EventController.patchFillWishlistCounterCtrl
-    )
     .patch("/edit/:eventId", doJwtAuth, EventController.patchEditEventCtrl)
     .patch(
-      "/toggleCancelled/:eventId",
-      doJwtAuth,
-      EventController.patchEventIsCancelledCtrl
+        "/toggleCancelled/:eventId",
+        doJwtAuth,
+        EventController.patchEventIsCancelledCtrl
     );
-
