+++ backend/src/services/index.js
@@ -10,9 +10,7 @@ import { getUserProfileInfo } from "./getUserProfileInfo.js";
 import { addEventToWishlist } from "./addEventToWishlist.js";
 import { editUserProfile } from "./editUserProfile.js";
 import { addEventToRegisteredEventsList } from "./addEventToRegisteredEventsList.js";
-import { fillRegisteredGuestsArray } from "./fillRegisteredGuestsArray.js";
 import { toggleIsCancelled } from "./toggleIsCancelled.js";
-import { fillEventWishlistCounter } from "./fillEventWishlistCounter.js";
 import { editEvent } from "./editEvent.js";
 import { addReview } from "./addReview.js";
 import { getHostProfileDetails } from "./getHostProfileDetails.js";
@@ -33,8 +31,6 @@ export const EventService = {
     addNewEvent,
     getAllEvents,
     getEventDetails,
-    fillRegisteredGuestsArray,
-    fillEventWishlistCounter,
     editEvent,
     toggleIsCancelled,
 };
